roles

+++ routes/web.php
@@ -16,28 +16,29 @@ use App\Http\Controllers\BookController as Book;
 */
 
 Route::get('/', function () {
-    return view('welcome');
+    return view('home');
 });
 
 Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 Route::prefix('category')->name('c_')->group(function () {
-    Route::get('/', [Category::class, 'index'])->name('index');
-    Route::get('/create', [Category::class, 'create'])->name('create');
-    Route::post('/create', [Category::class, 'store'])->name('store');
-    Route::get('/show/{category}', [Category::class, 'show'])->name('show');
-    Route::delete('/delete/{category}', [Category::class, 'destroy'])->name('delete');
-    Route::get('/edit/{category}', [Category::class, 'edit'])->name('edit');
-    Route::put('/edit/{category}', [Category::class, 'update'])->name('update');
-    Route::delete('/delete-books/{category}', [Category::class, 'destroyAll'])->name('delete_books');
+    Route::get('/', [Category::class, 'index'])->name('index')->middleware('gate:user');
+    Route::get('/create', [Category::class, 'create'])->name('create')->middleware('gate:admin');
+    Route::post('/create', [Category::class, 'store'])->name('store')->middleware('gate:admin');
+    Route::get('/show/{category}', [Category::class, 'show'])->name('show')->middleware('gate:user');
+    Route::delete('/delete/{category}', [Category::class, 'destroy'])->name('delete')->middleware('gate:admin');
+    Route::get('/edit/{category}', [Category::class, 'edit'])->name('edit')->middleware('gate:admin');
+    Route::put('/edit/{category}', [Category::class, 'update'])->name('update')->middleware('gate:admin');
+    Route::delete('/delete-books/{category}', [Category::class, 'destroyAll'])->name('delete_books')->middleware('gate:admin');
 });
 Route::prefix('book')->name('b_')->group(function () {
-    Route::get('/', [Book::class, 'index'])->name('index');
-    Route::get('/create', [Book::class, 'create'])->name('create');
-    Route::post('/create', [Book::class, 'store'])->name('store');
-    Route::get('/show/{book}', [Book::class, 'show'])->name('show');
-    Route::delete('/delete/{book}', [Book::class, 'destroy'])->name('delete');
-    Route::get('/edit/{book}', [Book::class, 'edit'])->name('edit');
-    Route::put('/edit/{book}', [Book::class, 'update'])->name('update');
+    Route::get('/', [Book::class, 'index'])->name('index')->middleware('gate:user');
+    Route::get('/create', [Book::class, 'create'])->name('create')->middleware('gate:admin');
+    Route::post('/create', [Book::class, 'store'])->name('store')->middleware('gate:admin');
+    Route::get('/show/{book}', [Book::class, 'show'])->name('show')->middleware('gate:user');
+    Route::delete('/delete/{book}', [Book::class, 'destroy'])->name('delete')->middleware('gate:admin');
+    Route::get('/edit/{book}', [Book::class, 'edit'])->name('edit')->middleware('gate:admin');
+    Route::put('/edit/{book}', [Book::class, 'update'])->name('update')->middleware('gate:admin');
 });
+
